SSR: Handle closing tags whose openers had an attribute directive (#144)
Some attribute directives (such as `wp-context`, see #143), need to be run not only on an opening tag (e.g. `<div wp-context="...">` but also on the matching closing one. (For `wp-context`, this is required to reset the context to the value that corresponds to the enclosing scope's context.)

This needs somewhat complex logic to keep track of nested tags. Consider e.g. the following example (used in this PR's unit test):

```html
<div>Example: 
	<div foo-test="abc">
		<img>
		<span>This is a test></span>
		<div>Here is a nested div</div>
	</div>
</div>
```

Here, we want to run the `foo-test` attribute directive processor on the closing `</div>` that matches the opening `<div foo-test="abc">` -- but not on any of the other two closing `</div>`s.

Rather than requiring individual directive processors to implement the required tracking of nested tags, we implement it in the `wp_process_directives` function.

To that end, this commit also moves it into a separate file and changes its signature a bit for better testability. Finally, we're adding a unit test to verify the behavior described above.